Rename Polly.Hsoting to Polly.Extensions
