+++ src/Polly.Extensions.Tests/Polly.Extensions.Tests.csproj
@@ -8,13 +8,13 @@
     <SkipPollyUsings>true</SkipPollyUsings>
     <Threshold>100</Threshold>
     <NoWarn>$(NoWarn);SA1600;SA1204</NoWarn>
-    <Include>[Polly.Hosting]*</Include>
+    <Include>[Polly.Extensions]*</Include>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Polly.Hosting\Polly.Hosting.csproj" />
+    <ProjectReference Include="..\Polly.Extensions\Polly.Extensions.csproj" />
   </ItemGroup>
 </Project>
