+++ src/Polly.Extensions/Polly.Extensions.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFrameworks>net7.0;net6.0;netstandard2.0;net472;net462</TargetFrameworks>
-    <AssemblyTitle>Polly.Hosting</AssemblyTitle>
-    <RootNamespace>Polly</RootNamespace>
+    <AssemblyTitle>Polly.Extensions</AssemblyTitle>
+    <RootNamespace>Polly.Extensions</RootNamespace>
     <Nullable>enable</Nullable>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <ProjectType>Library</ProjectType>
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <InternalsVisibleToTest Include="Polly.Hosting.Tests" />
+    <InternalsVisibleToTest Include="Polly.Extensions.Tests" />
   </ItemGroup>
 
   <ItemGroup>
